separate, rename
