+++ Assets/UniVRM-1.0/Components/BlendShape/PreviewMaterialItem.cs.meta
@@ -1,7 +1,5 @@
 fileFormatVersion: 2
-guid: 1c11db91c21a7e34095b5a32edd7d030
-timeCreated: 1523145842
-licenseType: Free
+guid: 90ee9936998e23342978b06befb3d9dd
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
