+++ Assets/UniVRM-1.0/Components/BlendShape/PreviewMeshItem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b634c6524bdf71744a1d448ccbec7eaa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
